pyproject.toml
[project]
name = "kitty-alert"
version = "0.1.0"
description = "A web app that alerts you when a new kitty is up for adoption from the SFPCA"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "django>=5.2.8",
  "django-debug-toolbar>=6.1.0",
  "django-extensions>=4.1",
  "django-phonenumber-field[phonenumberslite]>=7.2.0",
  "playwright>=1.48.0",
  "pylint-django>=2.6.1",
  "pytest-playwright>=0.7.1",
  "ruff>=0.14.4",
]
